Ensure we swap surface when region has changed.
We don't set content_changed to true if region has changed as
we omit the whole damage region calculations in this case.
Let's ensure content_changed is set true if region area has
changed, reset or surface has been cleared. This should ensure
we don't update the surface which is on screen.

Jira: None.
Test: No flicker seen when initialzing Video player on Linux and
      Android.

Signed-off-by: Kalyan Kondapally <kalyan.kondapally@intel.com>