Differentiate between physical disconnect and PowerMode state kOff.
In Disconnect, we where re-setting the power mode to off. This
had a counter effect when re-connecting the display as we want
to continue maintaining the same power mode as previous i.e. off,
doze etc. Let's not reset the power mode state tracker in
PhysicalDisplay, instead just call into DisplayQueue. This should
ensure we always power on to the correct power mode.

Jira: None.
Test: Enable HotPlugEvent traces and ensure we dont power off a
      re-connected display.
Signed-off-by: Kalyan Kondapally <kalyan.kondapally@intel.com>