Avoid full validation when adding new layers.
We can avoid re-validating whole commit when a layer is added
like we do for cursor. This is useful as complexity increases
with various combinations i.e. down/upscaling.

Also added tracing support to track down any duplicate layers
being added to a plane or in case we miss to add any layer.

Jira: None.
Test: No new regressions on Android.
Signed-off-by: Kalyan Kondapally <kalyan.kondapally@intel.com>