Track source correctly in displayplanestate.
We currently use scalars only for media but in future this
could be used for any layer. Let's make sure we track this
correctly and internally decide if source_rect should remain
the union of rects of all layers or equal to display frame.
This decision depends if display scalar flag is enabled
or disabled in displayplanestate.

Jira: None.
Test: Video still uses scaling as now.

Signed-off-by: Kalyan Kondapally <kalyan.kondapally@intel.com>