Weston: Don't use vblank for frame completion.
For some reason when using vblank callback as an indication
of frame completion the weston-simple-egl app reports a fps
of ~30-40. Thus instead of using the vblank callback, use the
received release fence.
This patch adds the release fence to the fds polled in the main
loop to use as an indication of frame completion.

Jira: None
Test: The weston-simple-egl app should report ~60 FPS

Signed-off-by: Harish Krupo <harish.krupo.kps@intel.com>