Add documentation for framebuffermanager.
This patch is meant to add documentation to the file framebuffermanager
in anticipation of the doxygen patch that will be added. If you have
any feedback or comments on the documentation in this patch feel free
to comment below. Hopefully we can add enough documentation throughout
the project that'll promote increased open source contributions.

Jira: GSE-1575
Test: doxygen Doxyfile and generated index.html

Signed-off-by: Richard Avelar richard.avelar@intel.com