Do not need translate when source crop is not (0, 0)
In Android, when the source crop is not (0, 0), the surface damage
is already translated to global display coordinate, no translation
is needed.
When the source crop is (0, 0), no scaling is needed, just transofrm
the coordiate according to the rotation scenario.

Change-Id: Ib4cd394290458a189c92fa668190107bb92dc49e
Tracked-On: OAM-72393
Tests: The HVAC, independent rotation and Mosaic mode work well
Signed-off-by: HeYue <yue.he@intel.com>