Zoom in or out the rendering damage for different display
Add the "zoom in and out" code back for connecting 2 display with different
resolution. In the bug case, 2K and 1080P display are connected. the Rect
should be enlarged for 2K when the rect is calculated with 1080p source
coordinates. and only when the damage region is set by SF, the damage will
be zoomed in/out.

Change-Id: I60cdd681895e8fa4907a89c2450c6da410ba208b
Tests: Verified on Celadon with 2 displays and mosaic on GP
Tracked-On: OAM-73581
Signed-off-by: Shaofeng Tang <shaofeng.tang@intel.com>